requirements.txt
discord.py
aiohttp
websockets
chardet
translate
